fix: FP8 output tensor must match input dtype
Fix RuntimeError when benchmarking FP8 GEMMs with out= parameter.

Problem:
RuntimeError: Expected out tensor to have dtype c10::Float8_e4m3fn,
but got c10::BFloat16 instead

Root cause:
- torch.matmul() with out= parameter requires output dtype to match input dtype
- When inputs are FP8, output must also be FP8 (not BF16)
- My previous assumption about automatic BF16 accumulation was incorrect

Fix:
```python
# Before (incorrect):
c_list = [torch.empty((m, n), dtype=init_dtype) for ...]  # BF16 output
# RuntimeError: dtype mismatch ❌

# After (correct):
c_list = [torch.empty((m, n), dtype=init_dtype).to(dtype) for ...]  # FP8 output
# torch.matmul(a_fp8, b_fp8, out=c_fp8) ✅
```

Key insight:
- PyTorch's torch.matmul API requires out= tensor dtype consistency
- Hardware accumulation (FP32) happens internally, but output is FP8
- If you want BF16 output, don't use out= parameter (let PyTorch allocate)

Example:
```python
# With out= parameter (dtype must match)
a = torch.randn(100, 100, dtype=torch.bfloat16).to(torch.float8_e4m3fn)
b = torch.randn(100, 100, dtype=torch.bfloat16).to(torch.float8_e4m3fn)
c = torch.empty(100, 100, dtype=torch.float8_e4m3fn)  # Must be FP8
torch.matmul(a, b, out=c)  # ✅ Works

# Without out= parameter (auto dtype selection)
c = torch.matmul(a, b)  # PyTorch decides output dtype
```

Testing:
```bash
# Now works without RuntimeError
python -m primus.tools.benchmark.dense_gemm_bench \
    --model Llama3.1_8B --dtype fp8_e4m3
```

Impact:
- ✅ Fixes FP8 GEMM benchmarking crash
- ✅ Maintains out= parameter for memory reuse (performance critical)
- ✅ All tensor dtypes now consistent (A: FP8, B: FP8, C: FP8)

+++ primus/tools/benchmark/gemm_bench.py
@@ -54,7 +54,10 @@ def profile_gemm(m, n, k, dtype, trans_a, trans_b, duration_s=10.0):
         b_list = [
             torch.randn(b_shape, device=device, dtype=init_dtype).to(dtype) for _ in range(num_rotations)
         ]
-        c_list = [torch.empty((m, n), device=device, dtype=init_dtype) for _ in range(num_rotations)]
+        # Output tensor must also be FP8 when using out= parameter with FP8 inputs
+        c_list = [
+            torch.empty((m, n), device=device, dtype=init_dtype).to(dtype) for _ in range(num_rotations)
+        ]
     else:
         a_list = [torch.randn(a_shape, device=device, dtype=dtype) for _ in range(num_rotations)]
         b_list = [torch.randn(b_shape, device=device, dtype=dtype) for _ in range(num_rotations)]
